Extract shared superclass for tests with parsing
